Move `PinCardView.State` to `PinEntryUi`
