+++ app/src/main/java/org/simple/clinic/security/pin/PinEntryUi.kt
@@ -1,9 +1,12 @@
 package org.simple.clinic.security.pin
 
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
 interface PinEntryUi {
 
   // Not yet migrated to Mobius
-  fun moveToState(state: PinEntryCardView.State)
+  fun moveToState(state: State)
   fun hideError()
   fun showIncorrectPinErrorForFirstAttempt()
   fun showIncorrectPinErrorOnSubsequentAttempts(remaining: Int)
@@ -11,4 +14,16 @@ interface PinEntryUi {
   fun clearPin()
   fun dispatchAuthenticatedCallback(enteredPin: String)
   fun setForgotButtonVisible(visible: Boolean)
+  
+  sealed class State: Parcelable {
+
+    @Parcelize
+    object PinEntry : State()
+
+    @Parcelize
+    object Progress : State()
+
+    @Parcelize
+    data class BruteForceLocked(val timeTillUnlock: TimerDuration) : State()
+  }
 }
